Add trellis tests
